Move trajectory dependency functions out of deprecated module
